feat(platform): add form helpers && address comments
