+++ apps/docs/src/app/platform/component-docs/platform-forms/checkbox/platform-checkbox-examples/platform-checkbox-a11y.component.html
@@ -1,18 +1,18 @@
 <h3>Checkbox A11y Examples using aria-label, aria-labelledby, aria-disabled and title</h3>
 <p id="checkboxLable">checkbox using aria-labelledby</p>
-<fdp-form-group [multiLayout]="true">
-    <fdp-form-field [id]="'a11y1'" [noLabelLayout]="true" rank="1">
-        <fdp-checkbox [name]="'checkboxA11y1'" [isBinary]="true" ariaLabel="checkbox using aria-label"
+<fdp-form-group multiLayout="true">
+    <fdp-form-field id="a11y1" [noLabelLayout]="true" rank="1">
+        <fdp-checkbox name="'checkboxA11y1'" [isBinary]="true" ariaLabel="checkbox using aria-label"
             [(ngModel)]="checkboxA11y1"></fdp-checkbox>
     </fdp-form-field>
 
-    <fdp-form-field [id]="'a11y2'" rank="2">
-        <fdp-checkbox [name]="'checkboxA11y2'" [isBinary]="true" ariaLabelledBy="checkboxLable"
+    <fdp-form-field id="a11y2" rank="2">
+        <fdp-checkbox name="'checkboxA11y2'" [isBinary]="true" ariaLabelledBy="checkboxLable"
             [(ngModel)]="checkboxA11y2"></fdp-checkbox>
     </fdp-form-field>
 
-    <fdp-form-field [id]="'a11y3'" [label]="'Disabled checkbox using aria-disabled'" rank="3">
-        <fdp-checkbox [name]="'checkboxA11y3'" [isBinary]="true" title="It is a disabled checkbox" disabled="true"
+    <fdp-form-field id="a11y3" label="Disabled checkbox using aria-disabled" rank="3">
+        <fdp-checkbox name="checkboxA11y3" [isBinary]="true" title="It is a disabled checkbox" disabled="true"
             [(ngModel)]="checkboxA11y3"></fdp-checkbox>
     </fdp-form-field>
 </fdp-form-group>
