+++ apps/docs/src/app/platform/component-docs/platform-forms/platform-form-container/platform-form-container-header/platform-form-container-header.component.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'app-form-container-header',
-    templateUrl: './platform-form-container-header.component.html',
-    styleUrls: ['./platform-form-container-header.component.scss']
+    templateUrl: './platform-form-container-header.component.html'
 })
 export class PlatformFormContainerHeaderComponent {}
